core/mffloader/gui: Moved DLL design to an instance-based approach (both core and loader functions now operate on previously created instances, which may exist in parallel)

+++ project_files/gui/Model/Light/DirectionalLightModel.cs
@@ -20,7 +20,7 @@ namespace gui.Model.Light
             return new DirectionalLightViewModel(models, this);
         }
 
-        public DirectionalLightModel(IntPtr handle) : base(handle)
+        public DirectionalLightModel(UInt32 handle) : base(handle)
         {
 
         }
